+++ eras/alonzo/impl/CHANGELOG.md
@@ -40,6 +40,7 @@
 * Remove `NoThunks (ContextError era)` constraint from `EraPlutusContext` class
 * Remove `NoThunks` deriving instance for `CollectError`
 * Make `AlonzoContextError` constructors lazy
+* Add `ToJSON` and `FromJSON` instances for `AlonzoScript era`
 
 ### `cddl`